+++ lib/src/main/java/org/asamk/signal/manager/syncStorage/StorageSyncModels.java
@@ -92,8 +92,11 @@ public final class StorageSyncModels {
 
     public static ContactRecord localToRemoteRecord(Recipient recipient, IdentityInfo identity) {
         final var address = recipient.getAddress();
+        final var aciPresent = address.aci().isPresent();
+        final var pniPresent = address.pni().isPresent();
+
         final var builder = SignalContactRecord.Companion.newBuilder(recipient.getStorageRecord())
-                .e164(address.number().orElse(""))
+                .e164(pniPresent ? address.number().orElse("") : "")
                 .username(address.username().orElse(""))
                 .profileKey(recipient.getProfileKey() == null
                         ? ByteString.EMPTY
@@ -126,7 +129,7 @@ public final class StorageSyncModels {
                     .archived(recipient.getContact().isArchived())
                     .hidden(recipient.getContact().isHidden());
         }
-        if (identity != null) {
+        if (identity != null && aciPresent) {
             builder.identityKey(ByteString.of(identity.getIdentityKey().serialize()))
                     .identityState(localToRemote(identity.getTrustLevel()));
         }